did global component for about & philosophy
